refactor(core): isolate src/core from React dependencies
