+++ src/core/types/css.ts
@@ -1,0 +1,3 @@
+import type {Properties} from 'csstype';
+
+export type CSSProperties = Properties;
